Implemented enabling or disabling of pumps

+++ Assignments/Assignments/GasStationComputer/FuelTankStation.cpp
@@ -27,7 +27,7 @@ void FuelTankStation::Initialize()
 
 double FuelTankStation::WithdrawGas(double amount, int gasType)
 {
-	int gas = 0;
+	double gas = 0;
 
 	mutex[gasType]->Wait();
